Implement Image.Composite and blend modes
Via [vips_composite][https://jcupitt.github.io/libvips/API/current/libvips-conversion.html#vips-composite]
